+++ packages/dm-core-plugins/src/form/components/NumberFieldWithoutArrows.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+import { TextField } from '@equinor/eds-core-react'
+
+export const NumberFieldWithoutArrows = styled(TextField)`
+  /* Chrome, Safari, Edge, Opera */
+  input::-webkit-outer-spin-button,
+  input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+  }
+
+  /* Firefox */
+  input[type=number] {
+    -moz-appearance: textfield;
+  }
+  `