+++ packages/dm-core-plugins/src/form/templates/ObjectModelContainedTemplate.tsx
@@ -43,7 +43,7 @@ export const ObjectModelContainedTemplate = (
           ))}
         {isDefined && !(onOpen && !uiAttribute?.showInline) && (
           <TooltipButton
-            title='Expand'
+            title={isExpanded ? 'Collapse' : 'Expand'}
             button-variant='ghost_icon'
             button-onClick={() => setIsExpanded(!isExpanded)}
             icon={isExpanded ? chevron_up : chevron_down}