+++ packages/dm-core-plugins/src/table/utils.ts
@@ -1,24 +1,4 @@
-import {
-  TGenericObject,
-  TTableConfig,
-  TTableRowItem,
-} from '@development-framework/dm-core'
-
-export function createItemsFromDocument(
-  document: TGenericObject | null
-): TTableRowItem[] {
-  return document
-    ? Object.values(document)?.map((data, index) => {
-        const id: string = crypto.randomUUID()
-        return {
-          data,
-          index,
-          key: id,
-          id,
-        }
-      })
-    : []
-}
+import { TTableConfig } from '@development-framework/dm-core'
 
 const defaultConfig: TTableConfig = {
   columns: [{ data: 'name', label: 'Name' }, { data: 'type' }],