+++ packages/dm-core/src/components/SortableList/SortableList.tsx
@@ -3,12 +3,12 @@ import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable'
 
 import { SortableListProps } from './types'
 
-export const SortableList = <T extends { id: string | number }>({
-  items,
-  children,
-}: SortableListProps<T>) => {
+export const SortableList = ({ items, children }: SortableListProps) => {
   return (
-    <SortableContext items={items} strategy={verticalListSortingStrategy}>
+    <SortableContext
+      items={items.map((item) => item.key)}
+      strategy={verticalListSortingStrategy}
+    >
       {children}
     </SortableContext>
   )