+++ packages/dm-core/src/components/SortableList/types.ts
@@ -1,11 +1,11 @@
 import { ReactNode } from 'react'
 
-export type SortableListProps<T extends { id: string | number }> = {
-  items: T[]
+export type SortableListProps = {
+  items: any[]
   children: ReactNode
 }
 
-export type SortableContextProps<T extends { id: string | number }> = {
+export type SortableContextProps<T extends { key: string | number }> = {
   items: T[]
   children: ReactNode
   onReorder: (reorderedItems: T[]) => void