+++ packages/dm-core/src/components/Table/TableRow/TableRowActions/TableRowActions.tsx
@@ -2,9 +2,10 @@ import React, { useState } from 'react'
 import { Button, Icon, Menu, Table } from '@equinor/eds-core-react'
 import { delete_to_trash, more_vertical } from '@equinor/eds-icons'
 import { TableRowActionsProps } from '../../types'
+import { DeleteSoftButton } from '../../../DeleteSoftButton'
 
 export function TableRowActions(props: TableRowActionsProps) {
-  const { editMode, itemKey, removeItem, deleteItem } = props
+  const { editMode, item, removeItem } = props
   const [isMenuOpen, setIsMenuOpen] = useState<boolean>(false)
   const [menuButtonAnchor, setMenuButtonAnchor] =
     useState<HTMLButtonElement | null>(null)
@@ -12,21 +13,18 @@ export function TableRowActions(props: TableRowActionsProps) {
   return (
     <Table.Cell style={{ textAlign: 'center' }}>
       {editMode ? (
-        <Button
-          aria-label='Remove row'
-          color='danger'
-          variant='ghost_icon'
-          onClick={removeItem}
-        >
-          <Icon data={delete_to_trash} aria-hidden />
-        </Button>
+        <DeleteSoftButton
+          onClick={() => removeItem(item, false)}
+          title={'Remove row'}
+          ariaLabel={'Remove row'}
+        />
       ) : (
         <>
           <Button
             aria-label='Row actions'
             aria-haspopup='true'
             aria-expanded={isMenuOpen}
-            aria-controls={`row-object-menu-${itemKey}`}
+            aria-controls={`row-object-menu-${item.key}`}
             variant='ghost_icon'
             onClick={() => setIsMenuOpen(true)}
             ref={setMenuButtonAnchor}
@@ -36,11 +34,11 @@ export function TableRowActions(props: TableRowActionsProps) {
           <Menu
             anchorEl={menuButtonAnchor}
             aria-labelledby='anchor-default'
-            id={`row-object-menu-${itemKey}`}
+            id={`row-object-menu-${item.key}`}
             onClose={() => setIsMenuOpen(false)}
             open={isMenuOpen}
           >
-            <Menu.Item onClick={() => deleteItem(itemKey)}>Delete</Menu.Item>
+            <Menu.Item onClick={() => removeItem(item, true)}>Delete</Menu.Item>
           </Menu>
         </>
       )}